:fire: Clean models (#763)
* :fire: Clean up user model interface

* :crystal_ball: Further model cleanup

* :fire: Update models according to interfaces

* :green_heart:

+++ Toggl.Multivac/Models/IProject.cs
@@ -26,7 +26,7 @@ namespace Toggl.Multivac.Models
 
         bool? AutoEstimates { get; }
 
-        int? EstimatedHours { get; }
+        long? EstimatedHours { get; }
 
         double? Rate { get; }
